+++ src/main/java/us/com/rclabs/app/model/User.java
@@ -4,6 +4,7 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 
@@ -12,6 +13,7 @@ import jakarta.validation.constraints.NotBlank;
  * Armazena informações como nome de usuário, senha e email.
  */
 @Entity
+@Table(name = "app_users") // Change table name to avoid Oracle reserved keyword
 public class User {
 
     @Id
