+++ src/main/resources/templates/index.html
@@ -1,74 +1,18 @@
 <!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<html xmlns:th="http://www.thymeleaf.org">
 
 <head>
     <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Dashboard - Item Management System</title>
-    <!-- External CSS -->
+    <title>Welcome</title>
     <link rel="stylesheet" th:href="@{/css/styles.css}">
-    <!-- Font Awesome for icons -->
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
-    <!-- React dependencies -->
-    <script src="https://unpkg.com/react@17/umd/react.development.js" crossorigin></script>
-    <script src="https://unpkg.com/react-dom@17/umd/react-dom.development.js" crossorigin></script>
-    <script src="https://unpkg.com/@babel/standalone/babel.min.js"></script>
 </head>
 
 <body>
     <div class="app-container">
-        <!-- Sidebar Menu with pre-added toggle button -->
-        <div class="sidebar">
-            <div class="sidebar-header">
-                <h3>Item Management</h3>
-            </div>
-            <ul class="sidebar-menu">
-                <li class="active"><a href="/"><i class="fas fa-list"></i> Items</a></li>
-                <li><a href="/items/new"><i class="fas fa-plus"></i> New Item</a></li>
-                <li><a href="/categories"><i class="fas fa-tags"></i> Categories</a></li>
-                <li><a href="/settings"><i class="fas fa-cog"></i> Settings</a></li>
-            </ul>
-            <!-- Menu toggle button directly in HTML -->
-            <div class="menu-toggle">
-                <i class="fas fa-bars"></i>
-            </div>
-        </div>
-
-        <!-- Main Content Area -->
-        <div class="main-content">
-            <!-- Breadcrumbs -->
-            <div class="breadcrumbs">
-                <span>Items</span>
-            </div>
-
-            <!-- Page Content -->
-            <div class="content-wrapper">
-                <div class="welcome-dashboard">
-                    <h1>Welcome to Item Management System</h1>
-                    <p>Select an option from the menu to get started</p>
-
-                    <!-- Dashboard Stats -->
-                    <div class="dashboard-stats">
-                        <div class="stat-card">
-                            <h3>Total Items</h3>
-                            <p class="stat-number">125</p>
-                        </div>
-                        <div class="stat-card">
-                            <h3>Categories</h3>
-                            <p class="stat-number">12</p>
-                        </div>
-                        <div class="stat-card">
-                            <h3>Recent Updates</h3>
-                            <p class="stat-number">24</p>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
+        <h1>Welcome to the Application</h1>
+        <p th:text="${message}">Default message</p>
+        <a href="/items">View Items</a>
     </div>
-
-    <!-- External JS -->
-    <script type="text/babel" th:src="@{/js/app.js}"></script>
 </body>
 
 </html>
