+++ src/main/resources/templates/item-new.html
@@ -1,0 +1,47 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+    layout:decorate="~{layout/default}">
+
+<head>
+    <title>New Item</title>
+</head>
+
+<body>
+    <div layout:fragment="content" th:with="activeTab='new-item'">
+        <!-- Breadcrumbs -->
+        <div class="breadcrumbs">
+            <a href="/">Items</a> / <span>New Item</span>
+        </div>
+
+        <!-- Page Content -->
+        <div class="content-wrapper">
+            <h1>Add New Item</h1>
+
+            <form th:action="@{/items/save}" th:object="${item}" method="post">
+                <div class="form-group">
+                    <label for="name">Name</label>
+                    <input type="text" id="name" th:field="*{name}" class="form-control" required />
+                    <small class="text-danger" th:if="${#fields.hasErrors('name')}" th:errors="*{name}"></small>
+                </div>
+
+                <div class="form-group">
+                    <label for="price">Price</label>
+                    <input type="number" id="price" th:field="*{price}" class="form-control" step="0.01" required />
+                    <small class="text-danger" th:if="${#fields.hasErrors('price')}" th:errors="*{price}"></small>
+                </div>
+
+                <div class="form-group">
+                    <label for="description">Description</label>
+                    <textarea id="description" th:field="*{description}" class="form-control" rows="3"></textarea>
+                </div>
+
+                <div class="form-group">
+                    <button type="submit" class="btn btn-primary">Save</button>
+                    <a href="/" class="btn btn-secondary">Cancel</a>
+                </div>
+            </form>
+        </div>
+    </div>
+</body>
+
+</html>
