+++ src/main/resources/templates/item.html
@@ -1,98 +1,63 @@
 <!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<html xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+    layout:decorate="~{layout/default}">
 
 <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Items - Item Management System</title>
-    <!-- External CSS -->
-    <link rel="stylesheet" th:href="@{/css/styles.css}">
-    <!-- Font Awesome for icons -->
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
-    <!-- React dependencies -->
-    <script src="https://unpkg.com/react@17/umd/react.development.js" crossorigin></script>
-    <script src="https://unpkg.com/react-dom@17/umd/react-dom.development.js" crossorigin></script>
-    <script src="https://unpkg.com/@babel/standalone/babel.min.js"></script>
+    <title>Items</title>
 </head>
 
 <body>
-    <div class="app-container">
-        <!-- Sidebar Menu with pre-added toggle button -->
-        <div class="sidebar">
-            <div class="sidebar-header">
-                <h3>Item Management</h3>
-            </div>
-            <ul class="sidebar-menu">
-                <li><a href="/"><i class="fas fa-list"></i> Items</a></li>
-                <li class="active"><a href="#"><i class="fas fa-search"></i> Item Details</a></li>
-                <li><a href="/items/new"><i class="fas fa-plus"></i> New Item</a></li>
-                <li><a href="/categories"><i class="fas fa-tags"></i> Categories</a></li>
-                <li><a href="/settings"><i class="fas fa-cog"></i> Settings</a></li>
-            </ul>
-            <!-- Menu toggle button directly in HTML -->
-            <div class="menu-toggle">
-                <i class="fas fa-bars"></i>
-            </div>
+    <div layout:fragment="content" th:with="activeTab='items'">
+        <!-- Breadcrumbs -->
+        <div class="breadcrumbs">
+            <span>Items</span>
         </div>
 
-        <!-- Main Content Area -->
-        <div class="main-content">
-            <!-- Breadcrumbs -->
-            <div class="breadcrumbs">
-                <a href="/">Items</a> / <span>Item Details</span>
+        <!-- Page Content -->
+        <div class="content-wrapper" id="root">
+            <h1>Item Management</h1>
+
+            <!-- Action Buttons -->
+            <div class="action-buttons">
+                <a href="/items/new" class="btn btn-primary"><i class="fas fa-plus"></i> Add New Item</a>
             </div>
 
-            <!-- Page Content -->
-            <div class="content-wrapper">
-                <h1>Item Management</h1>
-
-                <!-- Action Buttons -->
-                <div class="action-buttons">
-                    <button class="btn btn-primary"><i class="fas fa-plus"></i> Add New Item</button>
-                </div>
-
-                <!-- Items Table -->
-                <div class="items-table">
-                    <table>
-                        <thead>
-                            <tr>
-                                <th>ID</th>
-                                <th>Name</th>
-                                <th>Category</th>
-                                <th>Price</th>
-                                <th>Actions</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            <tr>
-                                <td>1</td>
-                                <td>Item One</td>
-                                <td>Electronics</td>
-                                <td>$299.99</td>
-                                <td>
-                                    <a href="#" class="action-btn edit"><i class="fas fa-edit"></i></a>
-                                    <a href="#" class="action-btn delete"><i class="fas fa-trash"></i></a>
-                                </td>
-                            </tr>
-                            <tr>
-                                <td>2</td>
-                                <td>Item Two</td>
-                                <td>Clothing</td>
-                                <td>$49.99</td>
-                                <td>
-                                    <a href="#" class="action-btn edit"><i class="fas fa-edit"></i></a>
-                                    <a href="#" class="action-btn delete"><i class="fas fa-trash"></i></a>
-                                </td>
-                            </tr>
-                        </tbody>
-                    </table>
-                </div>
+            <!-- Items Table -->
+            <div class="items-table">
+                <table>
+                    <thead>
+                        <tr>
+                            <th>ID</th>
+                            <th>Name</th>
+                            <th>Description</th>
+                            <th>Price</th>
+                            <th>Actions</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        <tr th:if="${items.empty}">
+                            <td colspan="5">No Items Available</td>
+                        </tr>
+                        <tr th:each="item : ${items}">
+                            <td th:text="${item.id}">1</td>
+                            <td th:text="${item.name}">Item Name</td>
+                            <td th:text="${item.description}">Description</td>
+                            <td th:text="${'$' + #numbers.formatDecimal(item.price, 1, 2)}">$0.00</td>
+                            <td>
+                                <a th:href="@{'/items/edit/' + ${item.id}}" class="action-btn edit"><i
+                                        class="fas fa-edit"></i></a>
+                                <a th:href="@{'/items/delete/' + ${item.id}}" class="action-btn delete"
+                                    onclick="return confirm('Are you sure you want to delete this item?');">
+                                    <i class="fas fa-trash"></i>
+                                </a>
+                            </td>
+                        </tr>
+                    </tbody>
+                </table>
             </div>
         </div>
     </div>
-
-    <!-- External JS -->
-    <script type="text/babel" th:src="@{/js/app.js}"></script>
+    <script th:src="@{/js/items.js}"></script>
 </body>
 
 </html>
