+++ src/main/resources/templates/items.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+    layout:decorate="~{layout/main}">
+
+<head>
+    <title>Items</title>
+</head>
+
+<body>
+    <div layout:fragment="content">
+        <h1>Items List</h1>
+        <div th:if="${items}" class="items-grid">
+            <div th:each="item : ${items}" class="item-card">
+                <h3 th:text="${item.name}">Item Name</h3>
+                <p th:text="${item.description}">Description</p>
+            </div>
+        </div>
+    </div>
+</body>
+
+</html>
