+++ src/main/resources/templates/layout/default.html
@@ -2,38 +2,32 @@
 <html xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
 
 <head>
-    <!-- 
-        Layout padrão utilizado por todas as páginas.
-        Fornece estrutura comum, incluindo menu de navegação e rodapé.
-     -->
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title layout:title-pattern="$CONTENT_TITLE - $LAYOUT_TITLE">App</title>
+    <title layout:title-pattern="$CONTENT_TITLE - Item Management System">Item Management System</title>
+    <!-- External CSS -->
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
-    <link rel="stylesheet" href="/css/styles.css">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
+    <link rel="stylesheet" th:href="@{/css/styles.css}">
 </head>
 
 <body>
-    <!-- Include the navigation bar -->
-    <div th:replace="fragments/navigation :: navbar"></div>
+    <div class="app-container">
+        <!-- Include sidebar with dynamic activeTab parameter passed from content templates -->
+        <div th:replace="~{fragments/sidebar :: sidebar(activeTab=${activeTab})}"></div>
 
-    <!-- Main content -->
-    <div class="container mt-4">
-        <div layout:fragment="content"></div>
-    </div>
-
-    <!-- Footer -->
-    <footer class="footer mt-auto py-3 bg-light">
-        <div class="container text-center">
-            <span class="text-muted">© 2025 Your App</span>
+        <!-- Main Content Area -->
+        <div class="main-content">
+            <!-- Content will be injected here -->
+            <div layout:fragment="content"></div>
         </div>
-    </footer>
+    </div>
 
     <!-- JavaScript dependencies -->
     <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
-    <script src="/js/app.js"></script>
+    <script th:src="@{/js/app.js}"></script>
 </body>
 
 </html>
