+++ src/main/resources/templates/layouts/main.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
+
+<head>
+    <meta charset="UTF-8">
+    <title>FDC APP</title>
+    <link rel="stylesheet" href="/css/styles.css">
+    <th:block layout:fragment="head"></th:block>
+</head>
+
+<body>
+    <div class="app-container">
+        <div class="sidebar">
+            <div class="sidebar-header">FDC APP</div>
+            <ul class="sidebar-menu">
+                <li th:classappend="${#strings.equals(#httpServletRequest.requestURI, '/') ? 'active' : ''}">
+                    <a href="/"><i class="fas fa-home"></i>Dashboard</a>
+                </li>
+                <li th:classappend="${#strings.equals(#httpServletRequest.requestURI, '/items') ? 'active' : ''}">
+                    <a href="/items"><i class="fas fa-box"></i>Items</a>
+                </li>
+                <li th:classappend="${#strings.equals(#httpServletRequest.requestURI, '/users') ? 'active' : ''}">
+                    <a href="/users"><i class="fas fa-users"></i>Users</a>
+                </li>
+            </ul>
+        </div>
+        <div class="main-content">
+            <main layout:fragment="content"></main>
+        </div>
+    </div>
+    <script src="/js/app.js"></script>
+</body>
+
+</html>
